remove ClusterClient and ClusterServer from the api
