+++ packages/api/src/cluster/index.ts
@@ -3,8 +3,6 @@ import { MemberEventType, MembershipEvent } from './MemberShip';
 import { ClusterEvent } from './ClusterEvent';
 import { Cluster, JoinCluster, ClusterOptions } from './Cluster';
 import { MembersMap, MembersPort, MembersData } from './MemberMap';
-import { CreateClusterClient, ClusterClient, ClusterClientOptions } from './ClusterClient';
-import { ClusterServer, ClusterServerOptions, CreateClusterServer } from './ClusterServer';
 
 export {
   ConnectWorkerEvent,
@@ -18,10 +16,4 @@ export {
   MembersMap,
   MembersPort,
   MembersData,
-  CreateClusterClient,
-  ClusterClient,
-  ClusterClientOptions,
-  ClusterServer,
-  ClusterServerOptions,
-  CreateClusterServer,
 };
